requirements.txt
# Python 3.11
python-dotenv~=1.0.0
Django~=4.1.7
djangorestframework~=3.14.0
celery~=5.2.7
redis~=4.5.4
psycopg2-binary~=2.9.6
pytest~=7.3.1
freezegun==1.0.0
djangorestframework-simplejwt~=5.2.2
black~=23.3.0
flake8~=6.0.0
isort~=5.12.0